Moves internal/kafka To pkg/kafka
