+++ .github/ISSUE_TEMPLATE/bug_report.yaml
@@ -127,7 +127,7 @@ body:
       - internal/filter
       - internal/k8sconfig
       - internal/k8stest
-      - internal/kafka
+      - pkg/kafka
       - internal/kubelet
       - internal/metadataproviders
       - internal/sharedcomponent
