+++ .github/ISSUE_TEMPLATE/other.yaml
@@ -121,7 +121,7 @@ body:
       - internal/filter
       - internal/k8sconfig
       - internal/k8stest
-      - internal/kafka
+      - pkg/kafka
       - internal/kubelet
       - internal/metadataproviders
       - internal/sharedcomponent
