+++ .github/ISSUE_TEMPLATE/unmaintained.yaml
@@ -126,7 +126,7 @@ body:
       - internal/filter
       - internal/k8sconfig
       - internal/k8stest
-      - internal/kafka
+      - pkg/kafka
       - internal/kubelet
       - internal/metadataproviders
       - internal/sharedcomponent
