+++ pkg/kafka/metadata.yaml
